feat: cashu stuff
